feat: add validation for category and item forms, including password checks

+++ controllers/CategoryController.js
@@ -1,4 +1,17 @@
 const db = require("../db/categoryQueries");
+require("dotenv").config();
+const checkPassword = require("../utils/checkPassword");
+const { body, validationResult, matchedData } = require("express-validator");
+
+const validateCategory = [
+  body("title")
+    .trim()
+    .matches(/^[A-Za-z\s]+$/)
+    .withMessage("Title must contain only letters and spaces.")
+    .notEmpty()
+    .withMessage("Title cannot be empty"),
+  body("password").notEmpty().withMessage("Password is required."),
+];
 
 async function getCategories(req, res) {
   const categories = await db.getAllCategories();
@@ -34,8 +47,27 @@ async function getCategoryEditForm(req, res) {
 
 async function postCategoryEditForm(req, res) {
   const id = req.params.id;
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    const category = await db.getCategoryById(id);
+    return res.status(400).render("category/categoryEditForm", {
+      category,
+      errors: errors.array(),
+      oldInput: req.body,
+    });
+  }
+  const { title } = matchedData(req);
+  const password = req.body.password;
+  if (!checkPassword(password)) {
+    const category = await db.getCategoryById(id);
+    const passwordError = { msg: "Invalid password" };
+    return res.status(400).render("category/categoryEditForm", {
+      category,
+      errors: [passwordError],
+      oldInput: req.body,
+    });
+  }
   try {
-    const title = req.body.title;
     await db.changeCategoryParams(title, id);
     res.redirect("/");
   } catch (err) {
@@ -58,6 +90,25 @@ async function getCategoryDeleteForm(req, res) {
 
 async function postCategoryDeleteForm(req, res) {
   const id = req.params.id;
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    const category = await db.getCategoryById(id);
+    return res.status(400).render("category/categoryDeleteForm", {
+      category,
+      errors: errors.array(),
+      oldInput: req.body,
+    });
+  }
+  const password = req.body.password; // Always available
+  if (!checkPassword(password)) {
+    const category = await db.getCategoryById(id);
+    const passwordError = { msg: "Invalid password" };
+    return res.status(400).render("category/categoryDeleteForm", {
+      category,
+      errors: [passwordError],
+      oldInput: req.body,
+    });
+  }
   try {
     await db.deleteCategory(id);
     res.redirect("/");
@@ -67,13 +118,31 @@ async function postCategoryDeleteForm(req, res) {
 }
 
 async function getCategoryCreateForm(req, res) {
-  res.render("category/categoryCreateForm");
+  res.render("category/categoryCreateForm", {
+    oldInput: null,
+    errors: [],
+  });
 }
 
 async function postCategoryCreateForm(req, res) {
-  const name = req.body.title;
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).render("category/categoryCreateForm", {
+      errors: errors.array(),
+      oldInput: req.body,
+    });
+  }
+  const title = req.body.title;
+  const password = req.body.password;
+  if (!checkPassword(password)) {
+    const passwordError = { msg: "Invalid password" };
+    return res.status(400).render("category/categoryCreateForm", {
+      errors: [passwordError],
+      oldInput: req.body,
+    });
+  }
   try {
-    await db.createCategory(name);
+    await db.createCategory(title);
     res.redirect("/");
   } catch (err) {
     res.status(500).send("Error fetching category create form");
@@ -88,5 +157,5 @@ module.exports = {
   getCategoryDeleteForm,
   postCategoryDeleteForm,
   getCategoryCreateForm,
-  postCategoryCreateForm
+  postCategoryCreateForm,
 };
